requirements.txt
streamlit==1.38.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
streamlit-option-menu==0.3.12